Samples cleanup
 - Move common properties to samples parent
 - Avoid explicit `ComponentScan` by moving all the classes to the same package
