+++ spring-cloud-stream-samples/double/pom.xml
@@ -7,7 +7,7 @@
 	<packaging>jar</packaging>
 
 	<name>spring-cloud-stream-sample-double</name>
-	<description>Demo project for Spring XD module</description>
+	<description>Demo project for Aggregate Builder</description>
 
 	<parent>
 		<groupId>org.springframework.cloud</groupId>
@@ -16,9 +16,7 @@
 	</parent>
 
 	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<start-class>demo.SinkApplication</start-class>
-		<java.version>1.8</java.version>
+		<start-class>demo.DoubleApplication</start-class>
 	</properties>
 
 	<dependencies>
