+++ spring-cloud-stream-samples/extended/pom.xml
@@ -6,7 +6,7 @@
 	<artifactId>spring-cloud-stream-sample-extended</artifactId>
 	<packaging>jar</packaging>
 	<name>spring-cloud-stream-sample-extended</name>
-	<description>Demo project for Spring XD module</description>
+	<description>Demo project for extended module using aggregate builder</description>
 
 	<parent>
 		<groupId>org.springframework.cloud</groupId>
@@ -15,9 +15,7 @@
 	</parent>
 
 	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<start-class>demo.SinkApplication</start-class>
-		<java.version>1.8</java.version>
+		<start-class>extended.ExtendedApplication</start-class>
 	</properties>
 
 	<dependencies>
