+++ spring-cloud-stream-samples/pom.xml
@@ -14,6 +14,10 @@
 	  <artifactId>spring-cloud-stream-parent</artifactId>
 	  <version>1.0.0.BUILD-SNAPSHOT</version>
 	</parent>
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>1.8</java.version>
+	</properties>
 	<modules>
 		<module>source</module>
 		<module>sink</module>
