+++ spring-cloud-stream-samples/sink/src/main/java/demo/SinkApplication.java
@@ -18,12 +18,8 @@ package demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-import sink.LogSink;
 
 @SpringBootApplication
-@ComponentScan(basePackageClasses=LogSink.class)
 public class SinkApplication {
 
 	public static void main(String[] args) {
