+++ spring-cloud-stream-samples/sink/src/test/java/demo/ModuleApplicationTests.java
@@ -31,8 +31,6 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
-import sink.LogSink;
-
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = SinkApplication.class)
 @WebAppConfiguration
