+++ spring-cloud-stream-samples/source/src/main/java/demo/SourceApplication.java
@@ -18,12 +18,8 @@ package demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-import source.TimeSource;
 
 @SpringBootApplication
-@ComponentScan(basePackageClasses=TimeSource.class)
 public class SourceApplication {
 
 	public static void main(String[] args) {
