+++ spring-cloud-stream-samples/source/src/main/java/demo/TimeSourceOptionsMetadata.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package source;
+package demo;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Pattern;
