+++ spring-cloud-stream-samples/transform/pom.xml
@@ -7,7 +7,7 @@
 	<packaging>jar</packaging>
 
 	<name>spring-cloud-stream-sample-transform</name>
-	<description>Demo project for Spring XD module</description>
+	<description>Demo project for transform module</description>
 
 	<parent>
 		<groupId>org.springframework.cloud</groupId>
@@ -16,9 +16,7 @@
 	</parent>
 
 	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<start-class>demo.TransformApplication</start-class>
-		<java.version>1.8</java.version>
 	</properties>
 
 	<dependencies>
