+++ spring-cloud-stream-samples/transform/src/main/java/demo/LoggingTransformer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package transform;
+package demo;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
