+++ spring-cloud-stream-samples/transform/src/main/java/demo/TransformApplication.java
@@ -18,12 +18,8 @@ package demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-import transform.LoggingTransformer;
 
 @SpringBootApplication
-@ComponentScan(basePackageClasses=LoggingTransformer.class)
 public class TransformApplication {
 
 	public static void main(String[] args) {
